Move parser regression tests to integration test file
